% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sir.R
\name{calc_sir}
\alias{calc_sir}
\title{Calculate the measure for uncertainty for ranks}
\usage{
calc_sir(ranking_prob_mat = NA, sucras = NA, treatment_names)
}
\arguments{
\item{ranking_prob_mat}{square matrix where the rows represent ranks and the columns treatments.}

\item{sucras}{a vector of sucra values}

\item{treatment_names}{a vector of treatment names. Should match the order of treatments in ranking_prob_mat or sucras}
}
\value{
The separation in ranking metric.
}
\description{
Calculate the measure for uncertainty for ranks
}
