% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_sir.R
\name{local_sir}
\alias{local_sir}
\title{Calculate the local SIR for a subset of treatments}
\usage{
local_sir(
  diffs = NA,
  ses = NA,
  samples = NA,
  trts = NULL,
  largerbetter,
  subset = NA,
  topx = NA,
  bottomx = NA,
  do_loo = F
)
}
\arguments{
\item{diffs}{Matrix of relative effects, must have column names that are the name of each treatment or trts must be specified}

\item{ses}{Matrix of estimated standard errors for relative effects, must have column names that are the name of each treatment or trts must be specified}

\item{samples}{a matrix or data.frame of MCMC samples, where rows are MCMC samples and columns are relative effects (relative to anchor) for treatments.
must have column names that are the name of each treatment or trts must be specified}

\item{trts}{optional; vector of treatment names matching order in diffs and sds}

\item{largerbetter}{Logical, T if larger outcomes are good. Defauls is T}

\item{subset}{A character vector of treatment names to consider as the set of competing treatments}

\item{topx}{Calculate the local SIR for the top x treatments, as ranked by global SUCRA/P-scores}

\item{bottomx}{Calculate the local SIR for the bottom x treatments, as ranked by global SUCRA/P-scores}

\item{do_loo}{Should a leave-one-out analysis of the subset be performed? Default is F}
}
\value{
A named list with two elements:
df - a data frame with the information used to draw the plot
plot - a ggplot object
}
\description{
Calculate the local SIR for a subset of treatments
}
