% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo_sir.R
\name{loo_sir}
\alias{loo_sir}
\title{Given a matrix of relative effects and standard errors, calculate the LOO SIR contribution}
\usage{
loo_sir(diffs = NA, ses = NA, rank_mat = NA, trt_names = NA, largerbetter)
}
\arguments{
\item{diffs}{Matrix of relative effects}

\item{ses}{Matrix of estimated standard errors for relative effects}

\item{trt_names}{optional; vector of treatment names matching order in diffs and sds}

\item{largerbetter}{Logical, T if larger outcomes are good. Defauls is T}
}
\value{
vector of LOO SIR contributions
}
\description{
Given a matrix of relative effects and standard errors, calculate the LOO SIR contribution
}
