% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_scores.R
\name{p_scores}
\alias{p_scores}
\title{Calculate p-scores from a set of relative effects and standard errors}
\usage{
p_scores(diffs, ses, largerbetter = T, trt_names = NA)
}
\arguments{
\item{diffs}{Matrix of relative effects}

\item{ses}{Matrix of estimated standard errors for relative effects}

\item{largerbetter}{Logical, T if larger outcomes are good. Defauls is T}

\item{trt_names}{optional; vector of treatment names matching order in diffs and sds}
}
\value{
named vector of P-scores
}
\description{
Calculate p-scores from a set of relative effects and standard errors
}
