% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cumul.sir.R
\name{plot.cumul.sir}
\alias{plot.cumul.sir}
\title{Plot results of leave-one-out method}
\usage{
\method{plot}{cumul.sir}(x, labels = FALSE, trt.trunc = 3, digits = 3, ...)
}
\arguments{
\item{x}{R object of class \code{sir}.}

\item{labels}{A logical indicating whether treatment names should be
shown in the plot.}

\item{trt.trunc}{Number of characters to keep for each treatment name if labels = T}

\item{digits}{Minimal number of significant digits for global SIR, see
\code{\link{print.default}}.}

\item{\dots}{Additional arguments (ignored).}
}
\value{
A ggplot2 object.
}
\description{
Plot results of cumulative method for separation in ranking (SIR) metric
}
\details{
Plot results of cumulative method for separation in ranking (SIR) metric
(Wigle et al., 2024).
}
\references{
Wigle A, ... (2024):
Separation In Ranking: A Metric for Quantifying Uncertainty in Treatment
Hierarchies in Network Meta-Analysis
}
\author{
Augustine Wigle \email{amhwigle@uwaterloo.ca},
  Guido Schwarzer \email{guido.schwarzer@uniklinik-freiburg.de}
}
