% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pscores.R
\name{pscores}
\alias{pscores}
\title{Calculate P-scores from a set of relative effects and standard errors}
\usage{
pscores(TE, seTE, small.values = "desirable", trts = NULL)
}
\arguments{
\item{TE}{Matrix of relative effects}

\item{seTE}{Matrix of estimated standard errors for relative effects}

\item{small.values}{A character string specifying whether small
outcome values indicate a beneficial (\code{"desirable"}) or
harmful (\code{"undesirable"}) effect, can be abbreviated.}

\item{trts}{optional; vector of treatment names matching order in TE and sds}
}
\value{
named vector of P-scores
}
\description{
Calculate P-scores from a set of relative effects and standard errors
}
