% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_mat.R
\name{rank_mat}
\alias{rank_mat}
\title{Calculate a ranking probabilities matrix from mcmc samples}
\usage{
rank_mat(samples, positive_good)
}
\arguments{
\item{samples}{a matrix or data.frame of MCMC samples, where rows are MCMC samples and columns are relative effects (relative to anchor) for treatments.
must have column names that are the name of each treatment.}
}
\description{
Calculate a ranking probabilities matrix from mcmc samples
}
