% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankMCMC.R
\name{rankMCMC}
\alias{rankMCMC}
\title{Calculate a ranking probabilities matrix from MCMC samples}
\usage{
rankMCMC(x, small.values = "desirable", trts = NULL)
}
\arguments{
\item{x}{a matrix or data.frame of MCMC samples, where rows are MCMC samples and columns are relative effects (relative to anchor) for treatments.
must have column names that are the name of each treatment.}

\item{small.values}{A character string specifying whether small
outcome values indicate a beneficial (\code{"desirable"}) or
harmful (\code{"undesirable"}) effect, can be abbreviated.}

\item{trts}{character vector of treatment names, optional if samples has column names}
}
\value{
A matrix of ranking probabilities where rows are treatments and columns are ranks
}
\description{
Calculate a ranking probabilities matrix from MCMC samples
}
