% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdist.R
\name{refdist}
\alias{refdist}
\title{Generate reference distribution for given network structure (with no multi-arm trials)}
\usage{
refdist(net, d, nsim)
}
\arguments{
\item{net}{A netmeta object which has the desired structure including tau2}

\item{d}{a vector of the desired relative effects, must be in the same order as the treatment ordering in net$X.matrix}

\item{nsim}{number of times to simulate from reference distribution}
}
\value{
A vector of the POTH values
}
\description{
Generate reference distribution for given network structure (with no multi-arm trials)
}
