% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir-package.R
\docType{package}
\name{sir-package}
\alias{sir-package}
\title{sir: Brief overview of methods and general hints}
\description{
R package \bold{sir} allows to calculate the separation in ranking (SIR)
metric to quantify the uncertainty in a treatment hierarchy in network
meta-analysis (Wigle et al., 2024).
}
\details{
R package \bold{sir} provides the following methods:
\itemize{
\item Calculate the separation in ranking metric (\code{\link{sir}})
\item Conduct leave-one-out analysis (\code{\link{loo.sir}})
}

Type \code{help(package = "sir")} for a listing of R functions
available in \bold{sir}.

Type \code{citation("sir")} on how to cite \bold{sir} in
publications.

The development version of \bold{sir} is available on GitHub
\url{https://github.com/augustinewigle/sir}.
}
\references{
Wigle A, ... (2024):
Separation In Ranking: A Metric for Quantifying Uncertainty in Treatment
Hierarchies in Network Meta-Analysis
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/augustinewigle/sir}
  \item Report bugs at \url{https://github.com/augustinewigle/sir/issues}
}

}
\author{
Augustine Wigle \email{amhwigle@uwaterloo.ca},
  Guido Schwarzer \email{guido.schwarzer@uniklinik-freiburg.de}
}
\keyword{package}
