% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sir.R
\name{sir}
\alias{sir}
\title{Calculate separation in ranking (SIR) metric}
\usage{
sir(x, trts = NULL)
}
\arguments{
\item{x}{A square matrix with rankings or a vector of SUCRA values
or P-scores (see Details).}

\item{trts}{An optional vector with treatment names. Must match the
order of treatments provided for argument \code{x}.}
}
\value{
A list with elements 'sucras' and 'sir'.
}
\description{
Separation in ranking (SIR) is a metric to quantify the uncertainty in
a treatment hierarchy in network neta-analysis
}
\details{
Argument \code{x} can be
\itemize{
 \item a square matrix where the rows represent ranks and the columns
   treatments,
 \item a vector with SUCRAs or P-scores.
}
}
\examples{
\dontrun{
# Make results reproducible
set.seed(1909)

# R package netmeta must be available
library("netmeta")
data(smokingcessation)
p1 <- pairwise(list(treat1, treat2, treat3),
               event = list(event1, event2, event3), n = list(n1, n2, n3),
               data = smokingcessation, sm = "OR")
net1 <- netmeta(p1)

rg1 <- rankogram(net1)
rg1$ranking.matrix.common
t(rg1$ranking.matrix.common)

netrank(rg1, random = FALSE)
sir(t(rg1$ranking.matrix.common))
sir(netrank(rg1)$ranking.common)

data(Senn2013)
net2 <- netmeta(TE, seTE, treat1.long, treat2.long, studlab,
                data = Senn2013, sm = "MD", random = FALSE, nchar.trts = 4)

rg2 <- rankogram(net2)
rg2$ranking.matrix.common
t(rg2$ranking.matrix.common)

netrank(rg2, random = FALSE)
sir(t(rg2$ranking.matrix.common))
sir(netrank(rg2)$ranking.common)
}

}
\references{
Wiggle A, ... (2024):
Separation In Ranking: A Metric for Quantifying Uncertainty in Treatment
Hierarchies in Network Meta-Analysis
}
\author{
Augustine Wigle \email{amhwigle@uwaterloo.ca},
  Guido Schwarzer \email{guido.schwarzer@uniklinik-freiburg.de}
}
