% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumul.sir.R
\name{cumul.poth}
\alias{cumul.poth}
\alias{cumul}
\title{Cumulative method for precision of treatment hierarchy (POTH) metric}
\usage{
\method{cumul}{poth}(x, sort = TRUE, ...)

cumul(x, ...)
}
\arguments{
\item{x}{An R object of class \code{poth}.}

\item{sort}{A logical indicating whether results should be sorted
by decreasing ranking metric.}

\item{\dots}{Additional arguments.}
}
\value{
A vector of POTH values for the top 2 to n treatments with additional class \code{cumul.poth}
}
\description{
Cumulative method for precision of treatment hierarchy (POTH) metric
}
