% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumul.sir.R
\name{cumul.sir}
\alias{cumul.sir}
\alias{cumul}
\title{Cumulative method for separation in ranking (SIR) metric}
\usage{
\method{cumul}{sir}(x, sort = TRUE, ...)

cumul(x, ...)
}
\arguments{
\item{x}{An R object of class \code{sir}.}

\item{sort}{A logical indicating whether results should be sorted
by decreasing ranking metric.}

\item{\dots}{Additional arguments.}

\item{digits}{Minimal number of significant digits, see
\code{\link{print.default}}.}
}
\value{
A vector of sir values for the top 2 to n treatments with additional class \code{cumul.sir}
}
\description{
Cumulative method for separation in ranking (SIR) metric
}
