% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo.R
\name{loo.poth}
\alias{loo.poth}
\alias{loo}
\alias{print.loo.poth}
\title{Leave-one-out method for precision of treatment hierarchy (POTH) metric}
\usage{
\method{loo}{poth}(x, ...)

loo(x, ...)

\method{print}{loo.poth}(x, digits = 3, ...)
}
\arguments{
\item{x}{An R object of class \code{poth}.}

\item{\dots}{Additional arguments.}

\item{digits}{Minimal number of significant digits, see
\code{\link{print.default}}.}
}
\value{
A data frame with additional class \code{loo.poth} and the following
  variables:
\item{trt}{Treatment names.}
\item{score}{Ranking metric (global).}
\item{rank}{Treatment rank (global).}
\item{residual}{Residual (global POTH minus leave-one-out POTH.}
\item{ratio}{Ratio of residual devided by absolute sum of residuals.}
}
\description{
Leave-one-out method for precision of treatment hierarchy (POTH) metric
}
\examples{
\dontrun{
# R package netmeta must be available
library("netmeta")
data(smokingcessation)
p1 <- pairwise(list(treat1, treat2, treat3),
  event = list(event1, event2, event3), n = list(n1, n2, n3),
  data = smokingcessation, sm = "OR")
net1 <- netmeta(p1, random = FALSE)

# Leave-one-out method
loo1 <- loo(poth(net1))
loo1

data(Senn2013)
net2 <- netmeta(TE, seTE, treat1.long, treat2.long, studlab,
                data = Senn2013, sm = "MD", random = FALSE)

}

}
\keyword{print}
