% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo.R
\name{loo.sir}
\alias{loo.sir}
\alias{loo}
\title{Leave-one-out method for separation in ranking (SIR) metric}
\usage{
\method{loo}{sir}(x, sort = TRUE, ...)

loo(x, ...)
}
\arguments{
\item{x}{An R object of class \code{sir}.}

\item{sort}{A logical indicating whether results should be sorted
by decreasing ranking metric.}

\item{\dots}{Additional arguments.}
}
\value{
A data frame with the following variables:
\item{residuals}{...}
\item{sir}{...}
\item{trt_name}{...}
\item{orig_score}{...}
\item{score_type}{...}
\item{rank}{...}
\item{ratio}{...}
}
\description{
Leave-one-out method for separation in ranking (SIR) metric
}
\examples{
\dontrun{
# R package netmeta must be available
library("netmeta")
data(smokingcessation)
p1 <- pairwise(list(treat1, treat2, treat3),
  event = list(event1, event2, event3), n = list(n1, n2, n3),
  data = smokingcessation, sm = "OR")
net1 <- netmeta(p1, random = FALSE)

# Use P-scores to calculate SIR
loo1 <- loo(sir(net1))
loo1
}

}
