% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo_sir.R
\name{loo_sir}
\alias{loo_sir}
\title{Given a matrix of relative effects and standard errors, calculate the LOO SIR contribution}
\usage{
loo_sir(
  diffs = NA,
  ses = NA,
  samples = NA,
  trts = NULL,
  largerbetter,
  plot_trts = TRUE
)
}
\arguments{
\item{diffs}{Matrix of relative effects, must have column names that are the name of each treatment or trts must be specified}

\item{ses}{Matrix of estimated standard errors for relative effects, must have column names that are the name of each treatment or trts must be specified}

\item{samples}{a matrix or data.frame of MCMC samples, where rows are MCMC samples and columns are relative effects (relative to anchor) for treatments.
must have column names that are the name of each treatment or trts must be specified}

\item{trts}{optional; vector of treatment names matching order in diffs and sds}

\item{largerbetter}{Logical, T if larger outcomes are good. Defauls is T}

\item{plot_trts}{Logical; should treatment names be included in the plot? Default it TRUE}
}
\value{
A named list with two elements:
df - a dataframe with the information used to draw the plot
plot - a ggplot object
}
\description{
Given a matrix of relative effects and standard errors, calculate the LOO SIR contribution
}
