% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.loo.sir.R
\name{plot.loo.sir}
\alias{plot.loo.sir}
\title{Plot results of leave-one-out method}
\usage{
\method{plot}{loo.sir}(x, plot_trts = TRUE, ...)
}
\arguments{
\item{x}{R object of class \code{sir}.}

\item{plot_trts}{A logical indicating whether treatment names be included
in the plot.}

\item{\dots}{Additional arguments (ignored).}
}
\value{
A ggplot2 object.
}
\description{
Plot results of leave-one-out method for separation in ranking (SIR) metric
}
\details{
Plot results of leave-one-out method for separation in ranking (SIR) metric
(Wiggle et al., 2024).
}
\examples{
\dontrun{
# R package netmeta must be available
library("netmeta")
data(smokingcessation)
p1 <- pairwise(list(treat1, treat2, treat3),
  event = list(event1, event2, event3), n = list(n1, n2, n3),
  data = smokingcessation, sm = "OR")
net1 <- netmeta(p1, random = FALSE)

# Use P-scores to calculate SIR
loo1 <- loo(sir(net1))
loo1
plot(loo1)
}

}
\references{
Wiggle A, ... (2024):
Separation In Ranking: A Metric for Quantifying Uncertainty in Treatment
Hierarchies in Network Meta-Analysis
}
\author{
Augustine Wigle \email{amhwigle@uwaterloo.ca},
  Guido Schwarzer \email{guido.schwarzer@uniklinik-freiburg.de}
}
