% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pscores.R
\name{pscores}
\alias{pscores}
\title{Calculate P-scores from a set of relative effects and standard errors}
\usage{
pscores(diffs, ses, largerbetter = TRUE, trts = NULL)
}
\arguments{
\item{diffs}{Matrix of relative effects}

\item{ses}{Matrix of estimated standard errors for relative effects}

\item{largerbetter}{Logical, T if larger outcomes are good. Defauls is T}

\item{trts}{optional; vector of treatment names matching order in diffs and sds}
}
\value{
named vector of P-scores
}
\description{
Calculate P-scores from a set of relative effects and standard errors
}
