% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_mat.R
\name{rank_mat}
\alias{rank_mat}
\title{Calculate a ranking probabilities matrix from MCMC samples}
\usage{
rank_mat(samples, largerbetter, trts = NA)
}
\arguments{
\item{samples}{a matrix or data.frame of MCMC samples, where rows are MCMC samples and columns are relative effects (relative to anchor) for treatments.
must have column names that are the name of each treatment.}

\item{largerbetter}{logical indicating if larger values indicate a more effective treatment}

\item{trts}{character vector of treatment names, optional if samples has column names}
}
\description{
Calculate a ranking probabilities matrix from MCMC samples
}
