% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sir.R
\name{sir}
\alias{sir}
\alias{print.sir}
\alias{summary.sir}
\alias{print.summary.sir}
\title{Calculate separation in ranking (SIR) metric}
\usage{
sir(x, pooled, trts = NULL)

\method{print}{sir}(x, sort = TRUE, digits = 3, ...)

\method{summary}{sir}(object, ...)

\method{print}{summary.sir}(x, sort = TRUE, digits = 3, ...)
}
\arguments{
\item{x}{Mandatory argument with suitable information on the treatment
hierarchy (see Details).}

\item{pooled}{A character string indicating whether the treatment hierarchy
is based on a common or random effects model. Either \code{"common"} or
\code{"random"}, can be abbreviated.}

\item{trts}{An optional vector with treatment names. Must match the
order of treatments provided for argument \code{x}.}

\item{sort}{A logical indicating whether printout should be sorted
by decreasing ranking metric.}

\item{digits}{Minimal number of significant digits, see
\code{\link{print.default}}.}

\item{\dots}{Additional arguments (ignored).}

\item{object}{An object of class \code{summary.sir}.}
}
\value{
An object of class \code{sir} with corresponding \code{print}
function. The object is a list containing the following components:
\item{sir}{Separation in ranking metric.}
\item{ranking}{A named numeric vector with rankings, i.e.,
  SUCRAs or P-scores.}
\item{ranking.matrix}{A square matrix with the probabilities
  for each possible rank (if information is available).}
\item{pooled}{As defined above.}
}
\description{
Separation in ranking (SIR) is a metric to quantify the uncertainty in
a treatment hierarchy in network meta-analysis
}
\details{
This function calculates the separation in ranking (SIR) metric to quantify 
the uncertainty in a treatment hierarchy in network meta-analysis
(Wiggle et al., 2024).
  
Argument \code{x} providing information on the treatment hierarchy is the
only mandatory argument. The following input formats can be provided:
\itemize{
 \item vector representing a ranking metric, i.e., SUCRAs or P-scores,
 \item square matrix with the probabilities for each possible rank
 (with treatments in rows and ranks in columns).
}
It is also possible to provide an R object created with
\code{\link[netmeta]{netrank}} (ranking metric) or
\code{\link[netmeta]{rankogram}} (probabilities for each possible rank)
from R package \bold{netmeta}.
}
\examples{
\dontrun{
# R package netmeta must be available
library("netmeta")
data(smokingcessation)
p1 <- pairwise(list(treat1, treat2, treat3),
  event = list(event1, event2, event3), n = list(n1, n2, n3),
  data = smokingcessation, sm = "OR")
net1 <- netmeta(p1, random = FALSE)

# Calculate probabilities for each possible rank
set.seed(1909) # make results reproducible
rg1 <- rankogram(net1)
rg1

# Calculate SIR
s1 <- sir(rg1)
s1

# Also print probabilities for each possible rank
summary(s1)

# Use SUCRAs to calculate SIR
nr1 <- netrank(rg1)
nr1
sir(nr1)
sir(nr1$ranking.common)

data(Senn2013)
net2 <- netmeta(TE, seTE, treat1.long, treat2.long, studlab,
                data = Senn2013, sm = "MD", random = FALSE)

# Use P-scores to calculate SIR
nr2 <- netrank(net2)
nr2
sir(nr2)
}

}
\references{
Wiggle A, ... (2024):
Separation In Ranking: A Metric for Quantifying Uncertainty in Treatment
Hierarchies in Network Meta-Analysis
}
\author{
Augustine Wigle \email{amhwigle@uwaterloo.ca},
  Guido Schwarzer \email{guido.schwarzer@uniklinik-freiburg.de}
}
\keyword{print}
\keyword{summary}
